// src/server/enums/Error-message.enum.ts
export enum ErrorMessages {
    INSUFFICIENT_DATA = 'Insufficient data entered by the user.',
    TIME_OUT_ERROR = 'Oops timed-out! Please retry.',
    INVALID_INPUT = 'Please enter a valid input.',
    DATABASE_CONNECTION_ERROR = 'Database might not be available at the moment, please try again in few moments.',
    BAD_REQUEST = 'Requested command is not found, please enter a valid command or try using "isc --help" for more information.',
    NOT_FOUND = 'No Record Found for the Queried data'
};