// src/server/enums/Error.enum.ts
export enum Errors {
    INSUFFICIENT_DATA = 'Insufficient data',
    TIME_OUT_ERROR = 'Timed-out!',
    INVALID_INPUT = 'Invalid input',
    DATABASE_CONNECTION_ERROR = 'Database Connection Error',
    DATABASE_QUERY_ERROR = 'Database Query Error',
    BAD_REQUEST = 'Bad Request Provided',
    NOT_FOUND = 'No Record Found'
};